requirements.txt
Flask
Flask-Cors
requests
cryptography
PyJWT
gunicorn
pytest
pytest-cov
flake8